kevinsaem_v31_dev/requirements.txt
# requirements.txt
# 케빈샘 AI코딩 예약 시스템 - Python 의존성

# 웹 프레임워크
fastapi==0.115.0
uvicorn[standard]==0.30.0

# 템플릿 엔진
jinja2==3.1.4

# DB 연결
pyodbc==5.1.0

# 데이터 검증
pydantic==2.9.0
pydantic-settings==2.5.0

# HTTP 클라이언트 (카카오 알림톡)
httpx==0.27.0

# JWT 인증
python-jose[cryptography]==3.3.0

# 폼 데이터 파싱
python-multipart==0.0.9

# 비동기 스케줄러 (선택)
# celery==5.4.0
# redis==5.0.0

# 에러 추적 (선택)
# sentry-sdk[fastapi]==2.13.0
